move code to module
